Toolchain refactoring.
* Split GlslParser out from GlslAnalyzer.
* Slim down the Parser and Analyzer interfaces.
