+++ src/commonTest/kotlin/baaahs/gl/Toolchain.kt
@@ -1,10 +1,3 @@
 package baaahs.gl
 
-import baaahs.gl.glsl.GlslAnalyzer
-import baaahs.gl.shader.OpenShader
-
 val testToolchain = RootToolchain(testPlugins())
-
-fun GlslAnalyzer.openShader(src: String): OpenShader {
-    return openShader(import(src))
-}
