+++ src/commonTest/kotlin/baaahs/gl/patch/GlslGenerationSpec.kt
@@ -44,7 +44,7 @@ object GlslGenerationSpec : Spek({
             """.trimIndent()
         }
         val glslAnalyzer by value { testToolchain.glslAnalyzer }
-        val mainShader by value { glslAnalyzer.import(shaderText) }
+        val mainShader by value { testToolchain.import(shaderText) }
         val mutablePatch by value { MutablePatch { } }
         val resultContentType by value { Color }
         val linkedPatch by value {
@@ -408,7 +408,7 @@ object GlslGenerationSpec : Spek({
             }
 
             val mainPaintShader by value {
-                glslAnalyzer.import(
+                testToolchain.import(
                     """
                         // Main Paint Shader
                         void main( void ) {
@@ -419,7 +419,7 @@ object GlslGenerationSpec : Spek({
             }
 
             val otherPaintShader by value {
-                glslAnalyzer.import(
+                testToolchain.import(
                     """
                         // Other Paint Shader
                         void mainImage( out vec4 fragColor, in vec2 fragCoord ) {
