+++ src/commonTest/kotlin/baaahs/gl/shader/OpenShaderSpec.kt
@@ -1,10 +1,10 @@
 package baaahs.gl.shader
 
 import baaahs.describe
-import baaahs.gl.glsl.GlslAnalyzer
 import baaahs.gl.glsl.GlslCode
 import baaahs.gl.override
-import baaahs.gl.testPlugins
+import baaahs.gl.testToolchain
+import baaahs.show.Shader
 import baaahs.toEqual
 import ch.tutteli.atrium.api.verbs.expect
 import org.spekframework.spek2.Spek
@@ -15,9 +15,9 @@ object OpenShaderSpec : Spek({
         val src by value {
             "void main(float intensity) { gl_FragColor = vec4(gl_FragCoord, 0., 1.); };"
         }
-        val plugins by value { testPlugins() }
-        val shaderAnalysis by value { GlslAnalyzer(plugins).analyze(src) }
-        val openShader by value { GlslAnalyzer(plugins).openShader(shaderAnalysis) }
+        val shader by value { Shader("Title", src) }
+        val shaderAnalysis by value { testToolchain.analyze(shader) }
+        val openShader by value { testToolchain.openShader(shaderAnalysis) }
         val invocationStatement by value {
             openShader.invocationGlsl(
                 GlslCode.Namespace("p"), "toResultVar",
