+++ src/jsMain/kotlin/baaahs/app/ui/ShaderPreview.kt
@@ -51,8 +51,7 @@ val ShaderPreview = xComponent<ShaderPreviewProps>("ShaderPreview") { props ->
 
         val shaderType = props.previewShaderBuilder?.openShader?.shaderType ?: run {
             // TODO: This is duplicating work that happens later in PreviewShaderBuilder, which is rotten.
-            appContext.toolchain.openShader(props.shader!!)
-                .shaderType
+            appContext.toolchain.openShader(props.shader!!).shaderType
         }
 
         val preview = if (shaderType == ProjectionShader) {
